+++ pii-manager/src/pii_manager/api/file.py
@@ -9,13 +9,16 @@ import gzip
 import bz2
 import lzma
 from itertools import zip_longest
+from pathlib import Path
 
 from typing import Dict, List, TextIO, Iterable, Optional, Union
 
+from pii_manager import PiiEnum
 from pii_manager.api import PiiManager
 from pii_manager.piientity import PiiEntity, piientity_asdict
-from pii_manager.helper.exception import PiiManagerException
+from pii_manager.helper.exception import PiiManagerException, InvArgException
 from pii_manager.helper.json import CustomJSONEncoder
+from pii_manager.helper.types import TYPE_STR_LIST
 
 TYPE_RESULT = Union[str, Iterable[PiiEntity]]
 
@@ -83,6 +86,38 @@ def print_tasks(proc: PiiManager, out: TextIO):
         print(f" {pii.name}  [country={country}]\n   ", doc, file=out)
 
 
+def read_taskfile(filename: str) -> List[Dict]:
+    """
+    Read a list of task descriptors from a JSON file
+    """
+    with open(filename, encoding='utf-8') as f:
+        try:
+            tasklist = json.load(f)
+            for td in tasklist:
+                td['pii'] = PiiEnum[td['pii']]
+            return tasklist
+        except json.JSONDecodeError as e:
+            raise InvArgException('invalid task spec file {}: {}',
+                                  filename, e)
+        except KeyError as e:
+            if str(e) == 'pii':
+                raise InvArgException("missing 'pii' field in task descriptor in {}", filename)
+            else:
+                raise InvArgException("cannot find PiiEnum element '{}' for task descriptor in {}", e, filename)
+        except Exception as e:
+            raise InvArgException("cannot read taskfile '{}': {}", filename, e)
+
+
+def add_taskfile(filename: TYPE_STR_LIST, proc: PiiManager):
+    """
+    Add all tasks defined in a JSON file (or several) to a processing object
+    """
+    if isinstance(filename, (str, Path)):
+        filename = [filename]
+    for name in filename:
+        tasklist = read_taskfile(name)
+        proc.add_tasks(tasklist)
+
 # ----------------------------------------------------------------------
 
 
@@ -93,6 +128,7 @@ def process_file(
     country: List[str] = None,
     tasks: List[str] = None,
     all_tasks: bool = False,
+    taskfile: TYPE_STR_LIST = None,
     split: str = "line",
     mode: str = "replace",
     template: str = None,
@@ -101,7 +137,7 @@ def process_file(
     show_stats: bool = False,
 ) -> Dict:
     """
-    Process PII tasks on a file
+    Process a number of PII tasks on a text file
     """
     # Create the object
     proc = PiiManager(
@@ -113,6 +149,8 @@ def process_file(
         template=template,
         debug=debug,
     )
+    if taskfile:
+        add_taskfile(taskfile, proc)
     if show_tasks:
         print_tasks(proc, sys.stderr)