+++ pii-manager/CHANGES.md
@@ -2,9 +2,13 @@ v. 0.5.0
  * new task list parsing code, adding a "full" format based on dicts, in
    addition to the previous "simplified" format based on tuples
  * refactored to allow more than one task for a given PII and country
- * added the capability to add task processors programmatically
+ * added the capability to add task descriptors programmatically
+ * added reading task descriptors from a JSON file
+ * context validation spec, for all three task implementation types
  * TASK_ANY split into LANG_ANY & COUNTRY_ANY
-
+ * PII detectors for international phone numbers, for en-any & es-any
+ * PII detector for IP addresses, language independent
+ 
 v. 0.4.0
  * PII GOV_ID task for es-ES and en-AU
  * PII EMAIL_ADDRESS task